added explicit conversions from double to int where needed (no more warning messages :) )
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@1036 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/Closest.h
@@ -189,7 +189,7 @@ Connect(Roadmap<CFG, WEIGHT>* _rm, Stat_Class& Stats,
   RoadmapGraph<CFG, WEIGHT>* pMap = _rm->m_pRoadmap;
   vector<CFG> vertices;
   pMap->GetVerticesData(vertices);
-  const int k = min(kclosest,vertices.size());
+  const int k = (int)min(kclosest,vertices.size());
   
   vector< pair<VID,VID> > kp;
   // Find k closest cfgs to each cfg in the roadmap
